refactor(FILE-1A): extract attachment replacement into its helper to meet the file-size budget
